Rename `GoodJob::Job` to `GoodJob::Execution`
- Also renames `GoodJob::CurrentExecution` to `GoodJob::CurrentThread`
- Updates Dashboard controllers and views/partials; puts the "Active Job ID" column first in the table view; wraps UUIDs with `<code>` tags
