+++ engine/app/controllers/good_job/active_jobs_controller.rb
@@ -2,8 +2,9 @@
 module GoodJob
   class ActiveJobsController < GoodJob::BaseController
     def show
-      @jobs = GoodJob::Job.where("serialized_params ->> 'job_id' = ?", params[:id])
-                          .order(Arel.sql("COALESCE(scheduled_at, created_at) DESC"))
+      @executions = GoodJob::Execution.active_job_id(params[:id])
+                                      .order(Arel.sql("COALESCE(scheduled_at, created_at) DESC"))
+      raise ActiveRecord::RecordNotFound if @executions.empty?
     end
   end
 end
