+++ engine/app/views/good_job/cron_schedules/index.html.erb
@@ -7,7 +7,7 @@
           <th>Configuration</th>
           <th>Class</th>
           <th>Description</th>
-          <th>Next execution</th>
+          <th>Next scheduled</th>
         </thead>
         <tbody>
           <% @cron_schedules.each do |job_key, job| %>
