+++ engine/app/views/good_job/dashboards/index.html.erb
@@ -7,7 +7,7 @@
     <div class='me-4'>
       <small>Filter by job class</small>
       <br>
-      <% @filter.job_classes.each do |name, count| %>
+      <% @filter.job_classes.each do |(name, count)| %>
         <% if params[:job_class] == name %>
           <%= link_to(root_path(@filter.to_params(job_class: nil)), class: 'btn btn-sm btn-outline-secondary active', role: "button", "aria-pressed": true)  do %>
             <%= name %> (<%= count %>)
@@ -37,18 +37,18 @@
   </div>
 </div>
 
-<% if @filter.jobs.present? %>
-  <%= render 'good_job/shared/jobs_table', jobs: @filter.jobs %>
+<% if @filter.executions.present? %>
+  <%= render 'good_job/shared/executions_table', executions: @filter.executions %>
 
   <nav aria-label="Job pagination" class="mt-3">
     <ul class="pagination">
       <li class="page-item">
         <%= link_to({ after_scheduled_at: (@filter.last.scheduled_at || @filter.last.created_at), after_id: @filter.last.id }, class: "page-link") do %>
-          Next jobs <span aria-hidden="true">&raquo;</span>
+          Older executions <span aria-hidden="true">&raquo;</span>
         <% end %>
       </li>
     </ul>
   </nav>
 <% else %>
-  <em>No jobs present.</em>
+  <em>No executions present.</em>
 <% end %>
