--- engine/app/views/good_job/shared/_jobs_table.erb
@@ -1,48 +1,0 @@
-<div class="card my-3">
-  <div class="table-responsive">
-    <table class="table card-table table-bordered table-hover table-sm mb-0">
-      <thead>
-        <tr>
-          <th>GoodJob ID</th>
-          <th>ActiveJob ID</th>
-          <th>Job Class</th>
-          <th>Queue</th>
-          <th>Scheduled At</th>
-          <th>Error</th>
-          <th>
-            ActiveJob Params&nbsp;
-            <%= tag.button "Toggle", type: "button", class: "btn btn-sm btn-outline-primary", role: "button",
-                           data: { bs_toggle: "collapse", bs_target: ".job-params" },
-                           aria: { expanded: false, controls: jobs.map { |job| "##{dom_id(job, "params")}" }.join(" ") }
-            %>
-          </th>
-          <th>Actions</th>
-        </tr>
-      </thead>
-      <tbody>
-        <% jobs.each do |job| %>
-          <tr id="<%= dom_id(job) %>">
-            <td><%= link_to job.id, active_job_path(job.serialized_params['job_id'], anchor: dom_id(job)) %></td>
-            <td><%= link_to job.serialized_params['job_id'], active_job_path(job.serialized_params['job_id']) %></td>
-            <td><%= job.serialized_params['job_class'] %></td>
-            <td><%= job.queue_name %></td>
-            <td><%= job.scheduled_at || job.created_at %></td>
-            <td class="text-break"><%= truncate(job.error, length: 1_000) %></td>
-            <td>
-              <%= tag.button "Preview", type: "button", class: "btn btn-sm btn-outline-primary", role: "button",
-                data: { bs_toggle: "collapse", bs_target: "##{dom_id(job, 'params')}" },
-                aria: { expanded: false, controls: dom_id(job, "params") }
-              %>
-              <%= tag.pre JSON.pretty_generate(job.serialized_params), id: dom_id(job, "params"), class: "collapse job-params" %>
-            </td>
-            <td>
-              <%= button_to job_path(job.id), method: :delete, class: "btn btn-sm btn-outline-danger", title: "Delete job" do %>
-                <%= render "good_job/shared/icons/trash" %>
-              <% end %>
-            </td>
-          </tr>
-        <% end %>
-      </tbody>
-    </table>
-  </div>
-</div>
