+++ engine/app/views/layouts/good_job/base.html.erb
@@ -23,7 +23,7 @@
       <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav me-auto">
           <li class="nav-item">
-            <%= link_to "All Jobs", root_path, class: ["nav-link", ("active" if current_page?(root_path))] %>
+            <%= link_to "All Executions", root_path, class: ["nav-link", ("active" if current_page?(root_path))] %>
           </li>
           <li class="nav-item">
             <%= link_to "Cron Schedules", cron_schedules_path, class: ["nav-link", ("active" if current_page?(cron_schedules_path))] %>
