+++ lib/good_job/current_thread.rb
@@ -4,7 +4,7 @@ require 'active_support/core_ext/module/attribute_accessors_per_thread'
 module GoodJob
   # Thread-local attributes for passing values from Instrumentation.
   # (Cannot use ActiveSupport::CurrentAttributes because ActiveJob resets it)
-  module CurrentExecution
+  module CurrentThread
     # @!attribute [rw] cron_key
     #   @!scope class
     #   Cron Key
@@ -23,24 +23,24 @@ module GoodJob
     #   @return [Exception, nil]
     thread_mattr_accessor :error_on_retry
 
-    # @!attribute [rw] good_job
+    # @!attribute [rw] executions
     #   @!scope class
-    #   Cron Key
-    #   @return [GoodJob::Job, nil]
-    thread_mattr_accessor :good_job
+    #   Execution
+    #   @return [GoodJob::Execution, nil]
+    thread_mattr_accessor :execution
 
     # Resets attributes
     # @return [void]
     def self.reset
       self.cron_key = nil
-      self.good_job = nil
+      self.execution = nil
       self.error_on_discard = nil
       self.error_on_retry = nil
     end
 
-    # @return [String] UUID of the currently executing GoodJob::Job
+    # @return [String] UUID of the currently executing GoodJob::Execution
     def self.active_job_id
-      good_job&.active_job_id
+      execution&.active_job_id
     end
 
     # @return [Integer] Current process ID
