+++ lib/good_job/railtie.rb
@@ -14,11 +14,11 @@ module GoodJob
 
     initializer "good_job.active_job_notifications" do
       ActiveSupport::Notifications.subscribe "enqueue_retry.active_job" do |event|
-        GoodJob::CurrentExecution.error_on_retry = event.payload[:error]
+        GoodJob::CurrentThread.error_on_retry = event.payload[:error]
       end
 
       ActiveSupport::Notifications.subscribe "discard.active_job" do |event|
-        GoodJob::CurrentExecution.error_on_discard = event.payload[:error]
+        GoodJob::CurrentThread.error_on_discard = event.payload[:error]
       end
     end
