+++ spec/lib/good_job/lockable_spec.rb
@@ -2,8 +2,8 @@
 require 'rails_helper'
 
 RSpec.describe GoodJob::Lockable do
-  let(:model_class) { GoodJob::Job }
-  let!(:job) { model_class.create(active_job_id: SecureRandom.uuid, queue_name: "default") }
+  let(:model_class) { GoodJob::Execution }
+  let!(:execution) { model_class.create(active_job_id: SecureRandom.uuid, queue_name: "default") }
 
   describe '.advisory_lock' do
     around do |example|
@@ -70,22 +70,22 @@ RSpec.describe GoodJob::Lockable do
     end
 
     it 'returns first row of the query with a lock' do
-      expect(job).not_to be_advisory_locked
-      result_job = model_class.advisory_lock.first
-      expect(result_job).to eq job
-      expect(job).to be_advisory_locked
+      expect(execution).not_to be_advisory_locked
+      result_execution = model_class.advisory_lock.first
+      expect(result_execution).to eq execution
+      expect(execution).to be_advisory_locked
 
-      job.advisory_unlock
+      execution.advisory_unlock
     end
 
     it 'can lock an alternative column' do
-      expect(job).not_to be_advisory_locked
-      result_job = model_class.advisory_lock(column: :queue_name).first
-      expect(result_job).to eq job
-      expect(job).to be_advisory_locked(key: "good_jobs-default")
-      expect(job).not_to be_advisory_locked # on default key
+      expect(execution).not_to be_advisory_locked
+      result_execution = model_class.advisory_lock(column: :queue_name).first
+      expect(result_execution).to eq execution
+      expect(execution).to be_advisory_locked(key: "good_jobs-default")
+      expect(execution).not_to be_advisory_locked # on default key
 
-      job.advisory_unlock(key: "good_jobs-default")
+      execution.advisory_unlock(key: "good_jobs-default")
     end
   end
 
@@ -95,7 +95,7 @@ RSpec.describe GoodJob::Lockable do
       model_class.limit(2).with_advisory_lock do |results|
         records = results
 
-        expect(records).to include(job)
+        expect(records).to include(execution)
         expect(records).to all be_advisory_locked
       end
 
@@ -116,73 +116,73 @@ RSpec.describe GoodJob::Lockable do
 
   describe '#advisory_lock' do
     it 'results in a locked record' do
-      job.advisory_lock!
-      expect(job.advisory_locked?).to be true
-      expect(job.owns_advisory_lock?).to be true
+      execution.advisory_lock!
+      expect(execution.advisory_locked?).to be true
+      expect(execution.owns_advisory_lock?).to be true
 
-      other_thread_owns_advisory_lock = Concurrent::Promises.future(job, &:owns_advisory_lock?).value!
+      other_thread_owns_advisory_lock = Concurrent::Promises.future(execution, &:owns_advisory_lock?).value!
       expect(other_thread_owns_advisory_lock).to be false
 
-      job.advisory_unlock
+      execution.advisory_unlock
     end
 
     it 'returns true or false if the lock is acquired' do
-      expect(job.advisory_lock).to eq true
+      expect(execution.advisory_lock).to eq true
 
-      expect(Concurrent::Promises.future(job, &:advisory_lock).value!).to eq false
+      expect(Concurrent::Promises.future(execution, &:advisory_lock).value!).to eq false
 
-      job.advisory_unlock
+      execution.advisory_unlock
     end
 
     it 'can lock alternative values' do
-      job.advisory_lock!(key: "alternative")
-      expect(job.advisory_locked?(key: "alternative")).to be true
-      expect(job.advisory_locked?).to be false
+      execution.advisory_lock!(key: "alternative")
+      expect(execution.advisory_locked?(key: "alternative")).to be true
+      expect(execution.advisory_locked?).to be false
 
-      job.advisory_unlock(key: "alternative")
+      execution.advisory_unlock(key: "alternative")
     end
 
     it 'can lock alternative postgres functions' do
-      job.advisory_lock!(function: "pg_advisory_lock")
-      expect(job.advisory_locked?).to be true
-      job.advisory_unlock
+      execution.advisory_lock!(function: "pg_advisory_lock")
+      expect(execution.advisory_locked?).to be true
+      execution.advisory_unlock
     end
   end
 
   describe '#advisory_unlock' do
     it 'unlocks the record' do
-      job.advisory_lock!
+      execution.advisory_lock!
 
       expect do
-        job.advisory_unlock
-      end.to change(job, :advisory_locked?).from(true).to(false)
+        execution.advisory_unlock
+      end.to change(execution, :advisory_locked?).from(true).to(false)
     end
 
     it 'unlocks the record only once' do
-      job.advisory_lock!
-      job.advisory_lock!
+      execution.advisory_lock!
+      execution.advisory_lock!
 
       expect do
-        job.advisory_unlock
-      end.not_to change(job, :advisory_locked?).from(true)
+        execution.advisory_unlock
+      end.not_to change(execution, :advisory_locked?).from(true)
 
-      job.advisory_unlock
+      execution.advisory_unlock
     end
 
     it 'unlocks the record even after the record is destroyed' do
-      job.advisory_lock!
-      job.destroy!
+      execution.advisory_lock!
+      execution.destroy!
 
       expect do
-        job.advisory_unlock
-      end.to change(job, :advisory_locked?).from(true).to(false)
+        execution.advisory_unlock
+      end.to change(execution, :advisory_locked?).from(true).to(false)
     end
 
     it 'returns true or false if the unlock operation is successful' do
-      job.advisory_lock
+      execution.advisory_lock
 
-      expect(Concurrent::Promises.future(job, &:advisory_unlock).value!).to eq false
-      expect(job.advisory_unlock).to eq true
+      expect(Concurrent::Promises.future(execution, &:advisory_unlock).value!).to eq false
+      expect(execution.advisory_unlock).to eq true
 
       unless RUBY_PLATFORM.include?('java')
         expect(POSTGRES_NOTICES.first).to include "you don't own a lock of type ExclusiveLock"
@@ -193,64 +193,64 @@ RSpec.describe GoodJob::Lockable do
 
   describe '#advisory_locked?' do
     it 'reflects whether the record is locked' do
-      expect(job.advisory_locked?).to eq false
-      job.advisory_lock
-      expect(job.advisory_locked?).to eq true
-      job.advisory_unlock
-      expect(job.advisory_locked?).to eq false
+      expect(execution.advisory_locked?).to eq false
+      execution.advisory_lock
+      expect(execution.advisory_locked?).to eq true
+      execution.advisory_unlock
+      expect(execution.advisory_locked?).to eq false
     end
 
-    it 'is accurate even if the job has been destroyed' do
-      job.advisory_lock
-      expect(job.advisory_locked?).to eq true
-      job.destroy!
-      expect(job.advisory_locked?).to eq true
-      job.advisory_unlock
-      expect(job.advisory_locked?).to eq false
+    it 'is accurate even if the execution has been destroyed' do
+      execution.advisory_lock
+      expect(execution.advisory_locked?).to eq true
+      execution.destroy!
+      expect(execution.advisory_locked?).to eq true
+      execution.advisory_unlock
+      expect(execution.advisory_locked?).to eq false
     end
   end
 
   describe '#advisory_unlock!' do
     it 'unlocks the record entirely' do
-      job.advisory_lock!
-      job.advisory_lock!
+      execution.advisory_lock!
+      execution.advisory_lock!
 
       expect do
-        job.advisory_unlock!
-      end.to change(job, :advisory_locked?).from(true).to(false)
+        execution.advisory_unlock!
+      end.to change(execution, :advisory_locked?).from(true).to(false)
     end
   end
 
   describe '.advisory_unlock_session' do
     it 'unlocks all locks in the session' do
-      job.advisory_lock!
+      execution.advisory_lock!
 
       model_class.advisory_unlock_session
 
-      expect(job.advisory_locked?).to eq false
+      expect(execution.advisory_locked?).to eq false
     end
   end
 
   describe 'create_with_advisory_lock' do
-    it 'causes the job to be saved and locked' do
-      job = model_class.new
-      job.create_with_advisory_lock = true
-      job.save!
+    it 'causes the execution to be saved and locked' do
+      execution = model_class.new
+      execution.create_with_advisory_lock = true
+      execution.save!
 
-      expect(job).to be_advisory_locked
+      expect(execution).to be_advisory_locked
 
-      job.advisory_unlock
+      execution.advisory_unlock
     end
   end
 
   it 'is lockable' do
     ActiveRecord::Base.clear_active_connections!
-    job.advisory_lock!
+    execution.advisory_lock!
 
     expect do
-      Concurrent::Promises.future(job, &:advisory_lock!).value!
+      Concurrent::Promises.future(execution, &:advisory_lock!).value!
     end.to raise_error GoodJob::Lockable::RecordAlreadyAdvisoryLockedError
 
-    job.advisory_unlock
+    execution.advisory_unlock
   end
 end
