+++ spec/lib/good_job_spec.rb
@@ -34,9 +34,9 @@ describe GoodJob do
   end
 
   describe '.cleanup_preserved_jobs' do
-    let!(:recent_job) { GoodJob::Job.create!(finished_at: 12.hours.ago) }
-    let!(:old_unfinished_job) { GoodJob::Job.create!(scheduled_at: 2.days.ago, finished_at: nil) }
-    let!(:old_finished_job) { GoodJob::Job.create!(finished_at: 36.hours.ago) }
+    let!(:recent_job) { GoodJob::Execution.create!(finished_at: 12.hours.ago) }
+    let!(:old_unfinished_job) { GoodJob::Execution.create!(scheduled_at: 2.days.ago, finished_at: nil) }
+    let!(:old_finished_job) { GoodJob::Execution.create!(finished_at: 36.hours.ago) }
 
     it 'deletes finished jobs' do
       deleted_jobs_count = described_class.cleanup_preserved_jobs
