+++ spec/test_app/app/jobs/cleanup_job.rb
@@ -2,9 +2,9 @@ class CleanupJob < ApplicationJob
   self.queue_name = :cleanup
 
   def perform(limit = 2_000)
-    earliest_job_to_preserve = GoodJob::Job.finished.order(created_at: :desc).limit(limit).last
+    earliest_job_to_preserve = GoodJob::Execution.finished.order(created_at: :desc).limit(limit).last
     return if earliest_job_to_preserve.blank?
 
-    GoodJob::Job.where("created_at < ?", earliest_job_to_preserve.created_at).delete_all
+    GoodJob::Execution.where("created_at < ?", earliest_job_to_preserve.created_at).delete_all
   end
 end
